adds react-router
