+++ package.json
@@ -22,7 +22,7 @@
     "react": "^0.14.8",
     "react-dom": "^0.14.8",
     "react-redux": "^4.4.0",
-    "react-router": "^2.0.0",
+    "react-router": "^2.8.1",
     "redux": "^3.3.1",
     "webpack": "^1.13.3"
   },
