+++ src/app/components/Main.jsx
@@ -1,0 +1,13 @@
+const React = require('react');
+
+var Main = React.createClass({
+  render() {
+    return (
+      <div>
+        {this.props.children}
+      </div>
+    )
+  }
+})
+
+module.exports = Main
