+++ src/app/config/routes.jsx
@@ -1,0 +1,2 @@
+const React = require('react');
+const ReactRouter
